Add Ember Basic Dropdown
