+++ tests/helpers/click-trigger.js
@@ -1,0 +1,24 @@
+import Ember from 'ember';
+export function nativeClick(selector, options = {}) {
+  let mousedown = new window.Event('mousedown', { bubbles: true, cancelable: true, view: window });
+  let mouseup = new window.Event('mouseup', { bubbles: true, cancelable: true, view: window });
+  let click = new window.Event('click', { bubbles: true, cancelable: true, view: window });
+  Object.keys(options).forEach(key => {
+    mousedown[key] = options[key];
+    mouseup[key] = options[key];
+    click[key] = options[key];
+  });
+  let element = document.querySelector(selector);
+  Ember.run(() => element.dispatchEvent(mousedown));
+  this.$(element).focus();
+  Ember.run(() => element.dispatchEvent(mouseup));
+  Ember.run(() => element.dispatchEvent(click));
+}
+
+export function clickTrigger(scope, options = {}) {
+  let selector = '.ember-basic-dropdown-trigger';
+  if (scope) {
+    selector = scope + ' ' + selector;
+  }
+  nativeClick(selector, options);
+}
